chore(auth): public exports
