+++ apps/frontend/src/features/auth/api/auth.queries.ts
@@ -1,77 +1,4 @@
-import {
-  queryOptions,
-  useMutation,
-  useQuery,
-  useQueryClient,
-} from "@tanstack/vue-query";
-import { useAuthStore } from "@/features/auth/model/auth.store";
-import {
-  getCurrentUser,
-  login as loginApi,
-  logout as logoutApi,
-  register as registerApi,
-} from "./auth.api";
-
-export const authKeys = {
+export const queryKeys = {
   all: ["auth"] as const,
-  me: () => [...authKeys.all, "me"] as const,
+  me: () => [...queryKeys.all, "me"] as const,
 };
-
-export function currentUserOptions(enabled: () => boolean = () => true) {
-  return queryOptions({
-    queryKey: authKeys.me(),
-    queryFn: getCurrentUser,
-    retry: false,
-    staleTime: 5 * 60 * 1000,
-    enabled,
-  });
-}
-
-export function useCurrentUser() {
-  const authStore = useAuthStore();
-
-  return useQuery(
-    currentUserOptions(
-      () => authStore.isInitialized && authStore.isAuthenticated,
-    ),
-  );
-}
-
-export function useLoginMutation() {
-  const queryClient = useQueryClient();
-  const authStore = useAuthStore();
-
-  return useMutation({
-    mutationFn: loginApi,
-    onSuccess: ({ token, user }) => {
-      authStore.setSession(token);
-      queryClient.setQueryData(authKeys.me(), user);
-    },
-  });
-}
-
-export function useRegisterMutation() {
-  const queryClient = useQueryClient();
-  const authStore = useAuthStore();
-
-  return useMutation({
-    mutationFn: registerApi,
-    onSuccess: ({ token, user }) => {
-      authStore.setSession(token);
-      queryClient.setQueryData(authKeys.me(), user);
-    },
-  });
-}
-
-export function useLogoutMutation() {
-  const queryClient = useQueryClient();
-  const authStore = useAuthStore();
-
-  return useMutation({
-    mutationFn: logoutApi,
-    onSettled: () => {
-      authStore.clearSession();
-      queryClient.clear();
-    },
-  });
-}
