+++ apps/frontend/src/features/auth/index.ts
@@ -1,0 +1,11 @@
+export {
+  authQueryKeys,
+  useCurrentUser,
+  useLoginMutation,
+  useLogoutMutation,
+  useRegisterMutation,
+} from "./api";
+
+export { useAuthStore } from "./model";
+
+export { AuthPageShell, SignedIn, SignedOut } from "./ui";
