+++ apps/frontend/src/features/auth/model/auth.session.ts
@@ -1,6 +1,6 @@
 import type { QueryClient } from "@tanstack/vue-query";
 import { refresh as refreshApi } from "../api/auth.api";
-import { authKeys } from "../api/auth.queries";
+import { queryKeys } from "../api/auth.queries";
 import type { useAuthStore } from "./auth.store";
 
 export function createAuthSession(params: {
@@ -14,12 +14,12 @@ export function createAuthSession(params: {
       const data = await refreshApi();
 
       authStore.setSession(data.token);
-      queryClient.setQueryData(authKeys.me(), data.user);
+      queryClient.setQueryData(queryKeys.me(), data.user);
 
       return { token: data.token };
     } catch {
       authStore.clearSession();
-      queryClient.removeQueries({ queryKey: authKeys.me() });
+      queryClient.removeQueries({ queryKey: queryKeys.me() });
       return null;
     } finally {
       if (options?.markInitialized) {
