+++ apps/frontend/src/features/auth/model/index.ts
@@ -1,0 +1,2 @@
+export { createAuthSession } from "./auth.session";
+export { useAuthStore } from "./auth.store";
