+++ apps/frontend/src/features/auth/ui/index.ts
@@ -1,0 +1,3 @@
+export { default as AuthPageShell } from "./PageShell.vue";
+export { default as SignedIn } from "./SignedIn.vue";
+export { default as SignedOut } from "./SignedOut.vue";
